# backend/docs/swagger.yaml
definitions:
  analytics.UserStats:
    properties:
      active:
        format: int64
        type: integer
      admins:
        format: int64
        type: integer
      inactive:
        format: int64
        type: integer
      newThisMonth:
        format: int64
        type: integer
      total:
        format: int64
        type: integer
    type: object
  auditlog.AuditLogListResponse:
    properties:
      limit:
        type: integer
      logs:
        items:
          $ref: '#/definitions/auditlog.ResponseDTO'
        type: array
      page:
        type: integer
      total:
        type: integer
    type: object
  auditlog.ResponseDTO:
    properties:
      action:
        type: string
      client_ip:
        type: string
      created_at:
        type: string
      description:
        type: string
      entity:
        type: string
      entity_id:
        type: string
      id:
        type: integer
      status:
        type: integer
      user_agent:
        type: string
      user_id:
        type: string
    type: object
  auth.LoginRequestDTO:
    properties:
      password:
        example: Password123
        type: string
      phone:
        example: "+996500500500"
        type: string
    required:
    - password
    - phone
    type: object
  auth.LoginResponseDTO:
    properties:
      access_token:
        type: string
      message:
        type: string
      refresh_token:
        type: string
    type: object
  auth.LogoutRequestDTO:
    properties:
      refresh_token:
        type: string
    type: object
  auth.RefreshTokenRequest:
    properties:
      refresh_token:
        example: refresh_token
        type: string
    type: object
  user.Role:
    enum:
    - superuser
    - admin
    - user
    type: string
    x-enum-varnames:
    - RoleSuperuser
    - RoleAdmin
    - RoleUser
  user.UserListResponse:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      users:
        items:
          $ref: '#/definitions/user.UserResponseDTO'
        type: array
    type: object
  user.UserRequestDTO:
    properties:
      first_name:
        type: string
      last_name:
        type: string
      middle_name:
        type: string
      password:
        example: Password123
        type: string
      phone:
        example: "+996500500500"
        type: string
      role:
        $ref: '#/definitions/user.Role'
      telegram:
        type: string
    required:
    - password
    - phone
    type: object
  user.UserResponseDTO:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      first_name:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      last_name:
        type: string
      middle_name:
        type: string
      phone:
        type: string
      photo:
        type: string
      role:
        $ref: '#/definitions/user.Role'
      telegram:
        type: string
      updated_at:
        type: string
    type: object
  user.UserSelfUpdateDTO:
    properties:
      first_name:
        type: string
      last_name:
        type: string
      middle_name:
        type: string
      password:
        type: string
      phone:
        example: "+996500500500"
        type: string
      telegram:
        type: string
    type: object
  user.UserUpdateDTO:
    properties:
      first_name:
        type: string
      is_active:
        type: boolean
      last_name:
        type: string
      middle_name:
        type: string
      password:
        type: string
      phone:
        example: "+996500500500"
        type: string
      role:
        type: string
      telegram:
        type: string
    type: object
host: localhost:8082
info:
  contact: {}
  description: API for managing users and authentication in the Auth Service application.
  title: AUTH SERVICE API
  version: "1.0"
paths:
  /api/v1/audit:
    get:
      operationId: getAuditLogs
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      - description: Filter by entity ID
        in: query
        name: entity_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auditlog.AuditLogListResponse'
      security:
      - BearerAuth: []
      summary: Get all user actions
      tags:
      - audit
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      operationId: authLogin
      parameters:
      - description: Login credentials
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/auth.LoginRequestDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.LoginResponseDTO'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Login user
      tags:
      - auth
  /api/v1/auth/logout:
    post:
      consumes:
      - application/json
      operationId: authLogout
      parameters:
      - description: Optional refresh token to invalidate
        in: body
        name: body
        schema:
          $ref: '#/definitions/auth.LogoutRequestDTO'
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Logout user
      tags:
      - auth
  /api/v1/auth/refresh:
    post:
      consumes:
      - application/json
      operationId: authRefresh
      parameters:
      - description: Refresh token
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/auth.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.LoginResponseDTO'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Refresh access token
      tags:
      - auth
  /api/v1/users:
    get:
      operationId: listUsers
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      - description: Search by name or phone
        in: query
        name: search
        type: string
      - description: Filter by active status
        in: query
        name: is_active
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.UserListResponse'
      security:
      - BearerAuth: []
      summary: Get all users with pagination
      tags:
      - users
    post:
      consumes:
      - application/json
      operationId: createUser
      parameters:
      - description: User data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/user.UserRequestDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/user.UserResponseDTO'
      security:
      - BearerAuth: []
      summary: Create user (admin only)
      tags:
      - users
  /api/v1/users/{id}:
    delete:
      operationId: deleteUser
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete user (admin only)
      tags:
      - users
    get:
      operationId: getUser
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.UserResponseDTO'
      security:
      - BearerAuth: []
      summary: Get user by ID
      tags:
      - users
    patch:
      consumes:
      - application/json
      operationId: updateUser
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/user.UserUpdateDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.UserResponseDTO'
      security:
      - BearerAuth: []
      summary: Update user (admin only)
      tags:
      - users
  /api/v1/users/me:
    get:
      operationId: getMe
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.UserResponseDTO'
      security:
      - BearerAuth: []
      summary: Get current user profile
      tags:
      - users
    put:
      consumes:
      - application/json
      operationId: updateMe
      parameters:
      - description: Fields to update
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/user.UserSelfUpdateDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.UserResponseDTO'
      security:
      - BearerAuth: []
      summary: Update own profile
      tags:
      - users
  /api/v1/users/phone/{phone}:
    get:
      operationId: getUserByPhone
      parameters:
      - description: User phone
        in: path
        name: phone
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.UserResponseDTO'
      security:
      - BearerAuth: []
      summary: Get user by phone
      tags:
      - users
  /api/v1/users/stats:
    get:
      operationId: getUserStats
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/analytics.UserStats'
      security:
      - BearerAuth: []
      summary: Get user statistics
      tags:
      - analytics
  /health:
    get:
      operationId: healthCheck
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "503":
          description: Service Unavailable
          schema:
            additionalProperties: true
            type: object
      summary: Health check
      tags:
      - health
securityDefinitions:
  BearerAuth:
    description: 'Enter: Bearer {your_jwt_token} (e.g., Bearer eyJhbGciOiJIUzI1NiIs...)'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"